restructured folders, moving db and services into src subfolder
