+++ src/pages/api/v1/auth/forgot-password/index.ts
@@ -1,5 +1,5 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { sendPasswordCode } from "@/server/service/forgotPasswordCodes";
+import { sendPasswordCode } from "@/services/forgotPasswordCodes";
 import { ApiError } from "@/types/exceptions";
 
 export default async function handler(
