+++ src/pages/api/v1/auth/login.ts
@@ -1,4 +1,4 @@
-import { validateLogin } from "@/server/service/auth";
+import { validateLogin } from "@/services/auth";
 import { ApiError } from "@/types/exceptions";
 import type { NextApiRequest, NextApiResponse } from "next";
