+++ src/pages/api/v1/pets/index.ts
@@ -1,5 +1,5 @@
-import { Pet } from "@/server/db/models";
-import { createPet } from "@/server/service/pets";
+import { Pet } from "@/db/models";
+import { createPet } from "@/services/pets";
 import { ApiError } from "@/types/exceptions";
 import type { NextApiRequest, NextApiResponse } from "next";
